fix : ExistsById 메소드 적용

+++ src/main/java/roomescape/service/TimeService.java
@@ -4,6 +4,8 @@ package roomescape.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import roomescape.advice.ErrorCode;
+import roomescape.advice.RoomEscapeException;
 import roomescape.model.Time;
 import roomescape.repository.TimeRepository;
 
@@ -24,6 +26,9 @@ public class TimeService {
     }
 
     public void deleteTime(Long id) {
+        if (!repository.exitsById(id)) { // existsById로 이름 수정 권장
+            throw new RoomEscapeException(ErrorCode.TIME_NOT_FOUND);
+        }
         repository.deleteById(id);
     }
